separate  query&mutation classes for earch app
