+++ base/schema.py
@@ -3,7 +3,7 @@ from graphene_django.forms.mutation import DjangoModelFormMutation
 from graphene_django import DjangoListField
 from .models import *
 from .forms import MemberCreationForm
-from  graphene import Field
+import   graphene 
 ##################################
 ################################## OBJECTS TYPES
 class MembersType(DjangoObjectType):
@@ -21,6 +21,21 @@ class TagType(DjangoObjectType):
 ########################################
 ######################################## Forms Mutations
 class MembersMutation(DjangoModelFormMutation):
-    member = Field(MembersType)
+    member =  graphene.Field(MembersType)
     class Meta:
         form_class = MemberCreationForm
+
+
+
+
+### main mutation
+class Mutation(graphene.ObjectType):
+    add_member = MembersMutation.Field()
+
+
+### main query
+class Query(graphene.ObjectType):
+    all_members = graphene.List(MembersType)
+    
+    def resolve_all_members(root, info):
+        return CustomUser.objects.all()
