+++ community/schema.py
@@ -1,5 +1,5 @@
 from graphene_django import DjangoObjectType
-
+import graphene
 from community.models import Community,CommunityOwner
 
 class CommunityType(DjangoObjectType):
@@ -12,4 +12,8 @@ class CommunityOwnerType(DjangoObjectType):
     class Meta :
         model = CommunityOwner
         fields = '__all__'
-        
+
+
+class Query(graphene.ObjectType):
+    all_communitys = graphene.List(CommunityType)
+    all_communitysOwners = graphene.List(CommunityOwnerType)
